+++ src/JsonApi/Schema/Error/Error.php
@@ -1,7 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace WoohooLabs\Yang\JsonApi\Schema;
+namespace WoohooLabs\Yang\JsonApi\Schema\Error;
+
+use WoohooLabs\Yang\JsonApi\Schema\Link\ErrorLinks;
 
 final class Error
 {