+++ src/JsonApi/Schema/Error/ErrorSource.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace WoohooLabs\Yang\JsonApi\Schema;
+namespace WoohooLabs\Yang\JsonApi\Schema\Error;
 
 final class ErrorSource
 {