+++ tests/JsonApi/Schema/Link/LinkTest.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace WoohooLabs\Yang\Tests\JsonApi\Schema;
+namespace WoohooLabs\Yang\Tests\JsonApi\Schema\Link;
 
 use PHPUnit\Framework\TestCase;
-use WoohooLabs\Yang\JsonApi\Schema\Link;
+use WoohooLabs\Yang\JsonApi\Schema\Link\Link;
 
 class LinkTest extends TestCase
 {