+++ tests/JsonApi/Schema/Link/LinksTest.php
@@ -1,12 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace WoohooLabs\Yang\Tests\JsonApi\Schema;
+namespace WoohooLabs\Yang\Tests\JsonApi\Schema\Link;
 
 use PHPUnit\Framework\TestCase;
 use WoohooLabs\Yang\JsonApi\Exception\LinkException;
-use WoohooLabs\Yang\JsonApi\Schema\Link;
-use WoohooLabs\Yang\JsonApi\Schema\Links;
+use WoohooLabs\Yang\JsonApi\Schema\Link\Link;
+use WoohooLabs\Yang\JsonApi\Schema\Link\Links;
 
 class LinksTest extends TestCase
 {